labelbox-api/requirements.txt
aniso8601==8.1.1
attrs==20.3.0
backoff==1.10.0
backports-datetime-fromisoformat==1.0.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Ext==0.1
Flask-RESTful==0.3.8
flask-restplus==0.13.0
google-api-core==1.25.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
labelbox==2.4.9
MarkupSafe==1.1.1
ndjson==0.3.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
rsa==4.7
six==1.15.0
urllib3==1.26.3
Werkzeug==0.16.1
